Update minizip.hbp
In this way we look for zlib.h relative to the current harbour setup. I used ADDONS directory that is usually present in an harbour installation, but may be missing in some specific cases... but also the src dir may be missing. So check for local zlib.h must be moved as the last.
Something is not clear to me about minizip+zlib.h